Final Solution with comments
I DID IT!!!!! YES!!!!  It took some time and I had to rewatch a number of courses to pull it all together.  The hardest part for me was the document.getElementById("quote-box").innerHTML = html;.  I really struggled with the second function until I saw the course that showed me how to pass the info to the html.  Hooray!!!